+++ CHANGELOG.md
@@ -5,6 +5,12 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 
 ## [Unreleased]
 
+## 0.2.1 - 2016-08-04
+### Added
+- `extension-crate-metrics.rb` Added metric format validation
+### Fixed
+- Reduce the verbosity of the retry logging
+
 ## 0.2.0 - 2016-07-03
 ### Added
 - `extension-crate-events.rb` Added buffering flushing before shutdown